# contracts/openapi/agent-public.json
{
  "components": {
    "schemas": {
...
      "GateEnvironment": {
        "enum": [
          "DEVELOPMENT",
          "STAGING",
          "PRODUCTION"
        ],
        "title": "GateEnvironment",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationalGateKey": {
        "enum": [
          "MODEL_CREDENTIALS",
          "MODEL_LIFECYCLE_CAPACITY",
          "NETWORK_ISOLATION",
          "DATA_PROCESSING_LOCATION",
          "SOURCE_CONNECTORS",
          "SOURCE_ACL",
          "DATA_CLASSIFICATION_DLP",
          "EVALUATION_DATASET",
          "RELEASE_APPROVAL",
          "ACTION_APPROVAL",
          "TENANT_KMS",
          "RETENTION_LEGAL_HOLD",
          "AUDIT_RESILIENCE"
        ],
        "title": "OperationalGateKey",
        "type": "string"
      },
...
      "OperationalGatePortfolioEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OperationalGatePortfolio"
          },
          "message": {
            "default": "DWAI-ON operational gates loaded.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "SUCCESS",
            "title": "Status",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "OperationalGatePortfolioEnvelope",
        "type": "object"
      },
      "OperationalGateProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/OperationalGateProblemCode"
          },
          "context": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "title": "Context",
            "type": "object"
          },
          "correlationId": {
            "title": "Correlationid",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "instance",
          "correlationId"
        ],
        "title": "OperationalGateProblem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DWP Agent Runtime",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/gates": {
      "get": {
        "operationId": "list_operational_gates_v1_admin_gates_get",
        "parameters": [
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GateEnvironment",
              "default": "PRODUCTION"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGatePortfolioEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Insufficient gate permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate workflow conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate store unavailable"
          }
        },
        "summary": "List Operational Gates",
        "tags": [
          "operational-gates"
        ]
      }
    },
    "/v1/admin/gates/bootstrap": {
      "post": {
        "operationId": "bootstrap_operational_gates_v1_admin_gates_bootstrap_post",
        "parameters": [
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GateEnvironment",
              "default": "PRODUCTION"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapOperationalGatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGatePortfolioEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Insufficient gate permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate workflow conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate store unavailable"
          }
        },
        "summary": "Bootstrap Operational Gates",
        "tags": [
          "operational-gates"
        ]
      }
    },
    "/v1/admin/gates/{gate_key}": {
      "get": {
        "operationId": "get_operational_gate_v1_admin_gates__gate_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "gate_key",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OperationalGateKey"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GateEnvironment",
              "default": "PRODUCTION"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateDetailEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Insufficient gate permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate workflow conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate store unavailable"
          }
        },
        "summary": "Get Operational Gate",
        "tags": [
          "operational-gates"
        ]
      },
      "patch": {
        "operationId": "configure_operational_gate_v1_admin_gates__gate_key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "gate_key",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OperationalGateKey"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GateEnvironment",
              "default": "PRODUCTION"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureOperationalGateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateDetailEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Insufficient gate permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate workflow conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate store unavailable"
          }
        },
        "summary": "Configure Operational Gate",
        "tags": [
          "operational-gates"
        ]
      }
    },
    "/v1/admin/gates/{gate_key}/decision": {
      "post": {
        "operationId": "decide_operational_gate_v1_admin_gates__gate_key__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "gate_key",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OperationalGateKey"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GateEnvironment",
              "default": "PRODUCTION"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideOperationalGateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateDetailEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Insufficient gate permission"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate workflow conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalGateProblem"
                }
              }
            },
            "description": "Gate store unavailable"
          }
        },
        "summary": "Decide Operational Gate",
        "tags": [
          "operational-gates"
        ]
      }
    },
...
  }
}